Make the can_set_mount_point hypervisor option configurable
As the mount point can only be set with Xen, many people might not want
to have the field for this visible. So we add an option for this, which
defaults to False.

Note that there's no schema migration since the new attribute is
optional.

See also https://bugs.launchpad.net/horizon/+bug/1255136